Revert "Feature/alt nav"
