+++ src/components/Shared/ChangeDao.js
@@ -1,40 +1,18 @@
 import React from 'react';
-import {
-  IconButton,
-  Link,
-  Tooltip,
-  useBreakpointValue,
-} from '@chakra-ui/react';
-import { RiMenu3Line, RiArrowDropDownLine } from 'react-icons/ri';
+import { Icon, Link } from '@chakra-ui/react';
+import { RiArrowDownSLine } from 'react-icons/ri';
 import DaoSwitcherModal from '../Modal/DaoSwitcherModal';
 import { useModals } from '../../contexts/PokemolContext';
 
 const ChangeDao = () => {
   const { modals, openModal } = useModals();
-  const changeIcon = useBreakpointValue({
-    base: <RiArrowDropDownLine />,
-    sm: <RiArrowDropDownLine />,
-    md: <RiArrowDropDownLine />,
-    lg: <RiMenu3Line />,
-  });
 
   return (
     <>
-      <Tooltip
-        label='Change DAO'
-        aria-label='Change DAO'
-        placement='right'
-        hasArrow
-      >
-        <IconButton
-          icon={changeIcon}
-          size='lg'
-          variant='ghost'
-          isRound='true'
-          as={Link}
-          onClick={() => openModal('changeDao')}
-        />
-      </Tooltip>
+      <Link fontSize='xs' onClick={() => openModal('changeDao')}>
+        Change Dao
+        <Icon as={RiArrowDownSLine} />
+      </Link>
 
       <DaoSwitcherModal isOpen={modals.changeDao} />
     </>
